#4399 rule status update hook
