+++ src/frontend/src/pages/RulesPage/components/RulesetGeneralView.tsx
@@ -1,13 +1,12 @@
 import React, { useMemo, useState } from 'react';
 import { Box, Paper, Table, TableBody, TableContainer, useTheme } from '@mui/material';
-import { Rule, RuleStatus, isLeadership } from 'shared';
+import { Rule, isLeadership } from 'shared';
 import RuleRow from '../RuleRow';
 import RuleStatusTag from './RuleStatusTag';
 import RuleContent from './RuleContent';
 import RuleStatusHistoryModal from './RuleStatusHistoryModal';
-import { useSetRuleStatus } from '../../../hooks/rules.hooks';
+import { useRuleStatusUpdate, useSetRuleStatus } from '../../../hooks/rules.hooks';
 import { useCurrentUser } from '../../../hooks/users.hooks';
-import { useToast } from '../../../hooks/toasts.hooks';
 import { compareRuleCodes } from '../../../utils/rules.utils';
 
 interface RulesetGeneralViewProps {
@@ -30,11 +29,8 @@ const RulesetGeneralView: React.FC<RulesetGeneralViewProps> = ({
   navigateToRule
 }) => {
   const theme = useTheme();
-  const toast = useToast();
   const user = useCurrentUser();
   const [historyModalRule, setHistoryModalRule] = useState<Rule | null>(null);
-  // the rule currently being written, so only its checkboxes disable
-  const [pendingRuleId, setPendingRuleId] = useState<string | null>(null);
 
   const canUpdateStatus = isLeadership(user.role);
 
@@ -45,23 +41,10 @@ const RulesetGeneralView: React.FC<RulesetGeneralViewProps> = ({
 
   // Status in general view is independent of any project
   const { mutateAsync: setStatus } = useSetRuleStatus(rulesetId);
+  const { pendingRuleId, updateStatus } = useRuleStatusUpdate((ruleId, status) => setStatus({ ruleId, status }));
 
   // Sort once by rule code so top-level rows render in a stable numeric order.
   const sortedTopLevelRules = useMemo(() => [...topLevelRules].sort(compareRuleCodes), [topLevelRules]);
-
-  const handleStatusChange = async (ruleId: string, status: RuleStatus) => {
-    setPendingRuleId(ruleId);
-    try {
-      await setStatus({ ruleId, status });
-      toast.success('Rule status updated successfully');
-    } catch (error) {
-      if (error instanceof Error) {
-        toast.error(error.message);
-      }
-    } finally {
-      setPendingRuleId(null);
-    }
-  };
 
   return (
     <Box>
@@ -79,7 +62,7 @@ const RulesetGeneralView: React.FC<RulesetGeneralViewProps> = ({
                   <RuleStatusTag
                     rule={r}
                     isLeaf={r.subRuleIds.length === 0}
-                    onStatusChange={canUpdateStatus ? (status) => handleStatusChange(r.ruleId, status) : undefined}
+                    onStatusChange={canUpdateStatus ? (status) => updateStatus(r.ruleId, status) : undefined}
                     disabled={pendingRuleId === r.ruleId}
                     onInfoClick={setHistoryModalRule}
                   />
